refactor(module): Refactor the cli to folder cli
The src folder is necessary for the c++ project. The cli is moved to the folder cli.
